authentication and privatisation of end points

+++ mycontacts-backend/controllers/contactController.js
@@ -2,16 +2,16 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc get all the contact
 // routes GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc create new contact
 // routes POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log(req.body);
@@ -25,12 +25,13 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 //@desc get individual  the contact
 // routes GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -43,13 +44,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update the contact
 // routes PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Not Authorized");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -60,13 +65,17 @@ const updateContact = asyncHandler(async (req, res) => {
 });
 //@desc Delete the contact
 // routes DELETE  /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("Not Authorized");
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
